// util/network.enum.ts
export enum Network {
    localhost = 'localhost',
    hardhat = 'hardhat',
    rinkeby = 'rinkeby',
    ropsten = 'ropsten',
    kovan = 'kovan',
    goerli = 'goerli',
    sepolia = 'sepolia',
    bscTestnet = 'bscTestnet',
    optimismTestnet = 'optimismTestnet',
    arbitrumTestnet = 'arbitrumTestnet',
    mumbai = 'mumbai',
    polygonZkEVMTestnet = 'polygonZkEVMTestnet',
    hecoTestnet = 'hecoTestnet',
    fantomTestnet = 'fantomTestnet',
    fuji = 'fuji',
    sokol = 'sokol',
    chiado = 'chiado',
    moonbaseAlpha = 'moonbaseAlpha',
    alfajores = 'alfajores',
    auroraTestnet = 'auroraTestnet',
    harmonyTestnet = 'harmonyTestnet',
    autobahnTestnet = 'autobahnTestnet',
    spark = 'spark',
    cronosTestnet = 'cronosTestnet',
    evmosTestnet = 'evmosTestnet',
    bobaTestnet = 'bobaTestnet',
    cantoTestnet = 'cantoTestnet',
    baseTestnet = 'baseTestnet',
    mantleTestnet = 'mantleTestnet',
    scrollTestnet = 'scrollTestnet',
    lineaTestnet = 'lineaTestnet',
    ethMain = 'ethMain',
    bscMain = 'bscMain',
    optimismMain = 'optimismMain',
    arbitrumMain = 'arbitrumMain',
    arbitrumNova = 'arbitrumNova',
    polygon = 'polygon',
    polygonZkEVMMain = 'polygonZkEVMMain',
    hecoMain = 'hecoMain',
    fantomMain = 'fantomMain',
    avalanche = 'avalanche',
    gnosis = 'gnosis',
    moonriver = 'moonriver',
    moonbeam = 'moonbeam',
    celo = 'celo',
    auroraMain = 'auroraMain',
    harmonyMain = 'harmonyMain',
    autobahn = 'autobahn',
    fuse = 'fuse',
    cronos = 'cronos',
    evmosMain = 'evmosMain',
    bobaMain = 'bobaMain',
    cantoMain = 'cantoMain'
}